fix(battlescanner): compare motion scanner timers as same-signedness ints
movementTicks[i] > TICKS_PER_SCANNER_UPDATE compared a signed
std::vector<int> element against an unsigned constant. Every value
actually stored today is non-negative, so this happened not to
misbehave in practice, but it is a latent trap: a signed value being
promoted to unsigned by the comparison would read as huge and positive
if it were ever negative, silently changing which branch runs. Casting
the threshold to int makes the comparison mean what it looks like it
means regardless of future changes to how movementTicks is populated.

No observable behaviour changes for players today, since the values
compared are already always in range where signed and unsigned
comparison agree.

+++ game/state/battle/battlescanner.cpp
@@ -45,7 +45,7 @@ void BattleScanner::update(GameState &state, unsigned int ticks)
 		updateTicksAccumulated -= TICKS_PER_SCANNER_UPDATE;
 		for (size_t i = 0; i < movementTicks.size(); i++)
 		{
-			if (movementTicks[i] > TICKS_PER_SCANNER_UPDATE)
+			if (movementTicks[i] > (int)TICKS_PER_SCANNER_UPDATE)
 			{
 				movementTicks[i] -= TICKS_PER_SCANNER_UPDATE;
 				changed = true;
